Organized the graphQL a bit
