+++ web/static/js/graphql/Content/index.js
@@ -1,0 +1,5 @@
+import * as exportM from './mutations';
+import * as exportQ from './queries';
+
+export const mutations = exportM;
+export const queries = exportQ;
